feat: add confirmation for delete row

+++ app/helpers/row_helper.rb
@@ -39,7 +39,7 @@ module RowHelper
   end
 
   def row_tag_actions(row)
-    remove = button_to([row.layer, row], method: :delete) do
+    remove = button_to([row.layer, row], method: :delete, form: {data: {"turbo-confirm": t("common.confirm")}}) do
       embedded_svg("remove_item.svg")
     end
     tag.td(id: dom_id(row, :last)) do
